requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.4.2
executing==2.0.1
extra-streamlit-components==0.1.60
fonttools==4.47.0
frozendict==2.3.10
frozenlist==1.4.1
future==0.15.2
genai==2.1.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.38.0
google-genai==1.8.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.28.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.6.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mplfinance==0.12.10b0
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.2
openai==0.27.10
opstrat==0.1.7
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.4
peewee==3.17.0
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
plotly-express==0.4.1
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.1
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
regex==2024.11.6
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.39.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.11
webencodings==0.5.1
websockets==15.0.1
yarl==1.9.4
yfinance==0.2.54
zipp==3.17.0
